+++ .dockerignore
@@ -15,3 +15,4 @@ modules/blockchain_interface/ethereum/build/*
 modules/Blockchain/Ethereum/build/
 hidden_service
 modules/Database/system.db
+logs
