+++ modules/Utilities.js
@@ -5,7 +5,6 @@ const pem = require('pem');
 const fs = require('fs');
 const moment = require('moment');
 const ipaddr = require('ipaddr.js');
-const winston = require('winston');
 const Storage = require('./Storage');
 const _ = require('lodash');
 const _u = require('underscore');
@@ -22,17 +21,9 @@ const sortedStringify = require('sorted-json-stringify');
 const mkdirp = require('mkdirp');
 const path = require('path');
 const rimraf = require('rimraf');
-require('winston-daily-rotate-file');
-
-const pjson = require('../package.json');
-const runtimeConfigJson = require('../config/config.json')[process.env.NODE_ENV];
-
-require('winston-loggly-bulk');
-
 
 class Utilities {
     constructor() {
-        this.getLogger();
     }
 
     /**
@@ -80,13 +71,6 @@ class Utilities {
         });
     }
 
-    formatFileLogs(args) {
-        const date = moment().format('D/MM/YYYY hh:mm:ss');
-        const msg = `${date} - ${args.level} - ${args.message} - \n${JSON.stringify(args.meta, null, 2)}`;
-        return msg;
-    }
-
-
     /**
      * Check if there is a new version of ot-node
      * @param {String} [options.repo] - Github repo name i.e. OriginTrail/ot-node.
@@ -110,171 +94,7 @@ class Utilities {
      * @returns {*} - log function
      */
     static getLogger() {
-        let logLevel = 'trace';
-        if (process.env.LOGS_LEVEL_DEBUG) {
-            logLevel = 'debug';
-        }
 
-        const customColors = {
-            trace: 'grey',
-            notify: 'green',
-            debug: 'yellow',
-            info: 'white',
-            warn: 'yellow',
-            important: 'magenta',
-            error: 'red',
-            api: 'cyan',
-            job: 'cyan',
-        };
-
-        try {
-            const transports =
-                [
-                    new (winston.transports.Console)({
-                        colorize: 'all',
-                        timestamp: true,
-                        formatter: this.formatFileLogs,
-                        prettyPrint: object => JSON.stringify(object),
-                        stderrLevels: [
-                            'trace',
-                            'notify',
-                            'debug',
-                            'info',
-                            'warn',
-                            'important',
-                            'error',
-                            'api',
-                            'job',
-                        ],
-                    }),
-                    new (winston.transports.DailyRotateFile)({
-                        filename: 'otnode-%DATE%.log',
-                        datePattern: 'YYYY-MM-DD-HH',
-                        zippedArchive: true,
-                        maxSize: '20m',
-                        maxFiles: '14d',
-                        json: false,
-                        formatter: this.formatFileLogs,
-                        dirname: `logs`,
-                    }),
-                ];
-
-            if (process.env.SEND_LOGS && parseInt(process.env.SEND_LOGS, 10)) {
-                transports.push(new (winston.transports.Loggly)({
-                    inputToken: 'abfd90ee-ced9-49c9-be1a-850316aaa306',
-                    subdomain: 'origintrail.loggly.com',
-                    tags: [process.env.NODE_ENV, runtimeConfigJson.network.id, pjson.version],
-                    json: true,
-                }));
-            }
-
-            const logger = new (winston.Logger)({
-                colors: customColors,
-                level: logLevel,
-                levels: {
-                    error: 0,
-                    important: 1,
-                    job: 2,
-                    api: 3,
-                    warn: 4,
-                    notify: 5,
-                    info: 6,
-                    trace: 7,
-                    debug: 8,
-                },
-                transports,
-            });
-            winston.addColors(customColors);
-
-            // Extend logger object to properly log 'Error' types
-            const origLog = logger.log;
-            logger.log = (level, msg) => {
-                if (msg instanceof Error) {
-                    // eslint-disable-next-line prefer-rest-params
-                    const args = Array.prototype.slice.call(arguments);
-                    args[1] = msg.stack;
-                    origLog.apply(logger, args);
-                } else {
-                    const transformed = Utilities.transformLog(level, msg);
-                    if (!transformed) {
-                        return;
-                    }
-                    origLog.apply(logger, [transformed.level, transformed.msg]);
-                }
-            };
-            return logger;
-        } catch (e) {
-            console.error('Failed to create logger', e);
-            process.exit(1);
-        }
-    }
-
-    /**
-     * Skips/Transforms third-party logs
-     * @return {*}
-     */
-    static transformLog(level, msg) {
-        if (process.env.LOGS_LEVEL_DEBUG) {
-            return {
-                level,
-                msg,
-            };
-        }
-        if (msg.startsWith('connection timed out')) {
-            return null;
-        }
-        if (msg.startsWith('negotiation error')) {
-            return null;
-        }
-        if (msg.includes('received late or invalid response')) {
-            return null;
-        }
-        if (msg.includes('error with remote connection')) {
-            return null;
-        }
-        if (msg.includes('remote connection encountered error')) {
-            return null;
-        }
-        if (msg.startsWith('updating peer profile')) {
-            return null;
-        }
-        if (msg.includes('client cannot service request at this time')) {
-            return null;
-        }
-        if (msg.includes('KADemlia error') && msg.includes('Message previously routed')) {
-            return null;
-        }
-        if (msg.includes('gateway timeout')) {
-            return null;
-        }
-        if (msg.startsWith('connect econnrefused')) {
-            return null;
-        }
-        if (msg.includes('unable to route to tunnel')) {
-            return null;
-        }
-        if (msg.includes('socket hang up')) {
-            return null;
-        }
-        if (msg.includes('getaddrinfo')) {
-            return null;
-        }
-        if (msg.includes('read econnreset')) {
-            return null;
-        }
-        if (msg.includes('connect etimedout')) {
-            return null;
-        }
-        if (msg.includes('connect ehostunreach')) {
-            return null;
-        }
-        if (msg.includes('ssl23_get_server_hello')) {
-            return null;
-        }
-        return {
-            level,
-            msg,
-        };
     }
 
     /**
