+++ package.json
@@ -84,6 +84,7 @@
     "bunyan": "^1.8.12",
     "camel-case": "^3.0.0",
     "check-dependencies": "^1.1.0",
+    "colors": "^1.3.2",
     "cucumber": "^5.0.1",
     "d3-format": "^1.3.0",
     "deasync-promise": "^1.0.1",
@@ -159,6 +160,7 @@
     "winston": "^2.4.3",
     "winston-daily-rotate-file": "^3.4.1",
     "winston-loggly-bulk": "^2.0.3",
+    "winston-papertrail": "^1.0.5",
     "xml2js": "^0.4.19"
   }
 }
