+++ testnet/.origintrail-noderc
@@ -1,0 +1,5 @@
+{
+    "node_wallet": "0x56406D8D0deb20d096138319aE9a1af58aDe5a7d",
+    "node_private_key": "d81d1d4690f1c10de544cca91d25c6f901258ce7031de71a21223e652e8d92a9",
+    "node_ip": "192.168.1.14"
+}
