+++ testnet/papertrail.yml
@@ -1,5 +1,4 @@
 files:
-  - /ot-node/node.log
   - /ot-node/complete-node.log
 destination:
   host: logs4.papertrailapp.com
